Fix all linker and compiler errors: codebase now builds cleanly
- Remove duplicate function definitions from EM monolithic files
  (werewolf.c, M2.c, kav_fight.c, warlock.c, demon.c, clan.c,
   act_move.c, olc.c) that were already split into dedicated Genesis
   files (wwauspice.c, wwbreed.c, wwtribe.c, random.c, stance.c,
   sorcerer.c, mage.c, disc.c, dragon.c, drow.c, flag.c)
- Fix mod_damcap signature: now long mod_damcap(ch, victim)
- Fix one_hit 4th arg: replace string literal with int 0 in elemental.c
- Fix Makefile link order: -lcrypt after object files (GCC requires this)
- Fix extra_descr_free duplicate global in olc.c
- Fix sphere_name duplicate array in warlock.c
- Remove do_answer stub (already in act_comm.c)
- Add stubs.c with implementations for trivia/quest/auction stubs,
  get_color, hurt_fighting, mod_damcap, and crash-log globals

+++ GodWars_Modern/src/act_info.c
@@ -37,7 +37,7 @@
 
 
 
-int     get_ratio       args( ( CHAR_DATA *ch ) );
+long    get_ratio       args( ( CHAR_DATA *ch ) );
 
 char *	const	where_name	[] =
 {
